Created TagsParams class
